Split price estimator option helpers
